chore: Working on moving to new transfer controller
