+++ app/views/projects/edit.html.erb
@@ -34,7 +34,7 @@
       <h2 class="text-2xl font-extrabold dark:text-white"><%= t(".advanced.title") %></h2>
 
       <%= viral_card(title: t(".advanced.transfer.title")) do %>
-        <%= render "edit_advanced_transfer" %>
+        <%= render "projects/transfer/edit_transfer" %>
       <% end %>
     <% end %>
