+++ app/views/projects/edit.turbo_stream.erb
@@ -1,5 +1,5 @@
 <%= turbo_stream.update("edit_advanced_transfer") do %>
-  <%= render "transfer_form" %>
+  <%= render "form" %>
 <% end %>
 
 <%= turbo_stream.replace "errors", viral_alert(type:, message:) %>
