+++ app/views/projects/transfer/_edit_transfer.html.erb
@@ -19,5 +19,5 @@
 </ul>
 
 <%= turbo_frame_tag("edit_advanced_transfer") do %>
-  <%= render "transfer_form", project: @project %>
+  <%= render "projects/transfer/form", project: @project %>
 <% end %>
