+++ app/views/projects/transfer/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(url: namespace_project_transfer_path, method: :post, class: "grid gap-4") do |form| %>
+<%= form_with(url: namespace_project_transfer_index_path, method: :post, class: "grid gap-4") do |form| %>
   <%= form.label :new_namespace_id,
              t(:"projects.edit.advanced.transfer.new_namespace_id") %>
   <div class="form-field">
